+++ be/src/pipeline/exec/join/join_op.h
@@ -43,7 +43,7 @@ namespace doris::pipeline {
  *  RowRefListWithFlags is a list of many RowRefWithFlags. This means each row will have different visited flags. It's used for join operation which has `other_conjuncts`.
  */
 struct RowRef {
-    uint32_t row_num = 0;
+    size_t row_num = 0;
 
     RowRef() = default;
     RowRef(size_t row_num_count) : row_num(row_num_count) {}
